requirements-dev.txt
-r requirements.txt

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality
black>=24.0.0
mypy>=1.8.0
ruff>=0.1.0

# Type stubs
types-requests>=2.31.0
lxml-stubs>=0.5.0
